Fix an issue where the AmountOff percentage is cast to an int, even t… (#2004)
The admin panel allows you to enter a decimal value when creating an
AmountOff Discount with a percentage discount. However, it gets cast to
an int in the AmountOff class, which results in the percentage being
rounded and the discount amount being slightly off.

You could argue that usually percentage discounts are given as whole
numbers, but since the admin panel already supports decimal values, I
think this should be changed.

Fixes #2003

---------

+++ packages/core/src/DiscountTypes/AmountOff.php
@@ -224,7 +224,7 @@ class AmountOff extends AbstractDiscountType
     /**
      * Apply the percentage to the cart line.
      */
-    private function applyPercentage(int $value, Cart $cart): Cart
+    private function applyPercentage(float $value, Cart $cart): Cart
     {
         $lines = $this->getEligibleLines($cart);
